pyproject.toml
[project]
name = "group2-data-viz"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
]
